backend/requirements.txt
Django==4.0.2
django_filter==21.1
djangorestframework==3.13.1
